Remove pixel encoding module and related functions; refactor tile_range to include bitmap utility functions
